Completed JWT

+++ controllers/api/v1/posts_api.js
@@ -24,30 +24,31 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id) {
-        post.remove();
+        if (post.user == req.user.id) {
+            post.remove();
 
-        await Comment.deleteMany({ post: req.params.id });
+            await Comment.deleteMany({ post: req.params.id });
 
 
-        // if (req.xhr) {
-        //     return res.status(200).json({
-        //         data: {
-        //             post_id: req.params.id
-        //         },
-        //         message: "Post deleted"
-        //     });
-        // }
+            // if (req.xhr) {
+            //     return res.status(200).json({
+            //         data: {
+            //             post_id: req.params.id
+            //         },
+            //         message: "Post deleted"
+            //     });
+            // }
 
-        // req.flash('success', 'Post and associated comments deleted!');
+            // req.flash('success', 'Post and associated comments deleted!');
 
-        return res.json(200, {
-            message: "Post and Associated Comments are Deleted"
-        });
-        // } else {
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+            return res.json(200, {
+                message: "Post and Associated Comments are Deleted"
+            });
+        } else {
+            return res.json(401, {
+                message: "User cannot delete this post"
+            });
+        }
 
     } catch (err) {
         console.log("********", err);
